Create Android Chrome favicon parameters This function helps with the creation of the list structure required to provide Android Chrome-specific favicon parameters to the API. Further details can be found in the realfavicongenerator.net API documentation

Create Android Chrome favicon parameters This function helps with the creation of the list structure required to provide Android Chrome-specific favicon parameters to the API. Further details can be found in the realfavicongenerator.net API documentation

android_chrome_helper(
  picture_aspect = c("no_change", "background_and_margin", "shadow"),
  margin = NULL,
  background_color = NULL,
  manifest = list(name = "", start_url = NULL, display = c("browser", "standalone"),
    orientation = c("portrait", "landscape"), theme_color = NULL, existing_manifest =
    NULL, on_conflict = c("raise_error", "override", "keep_existing")),
  assets = list(legacy_icon = FALSE, low_resolution_icons = FALSE)
)

Arguments

picture_aspect
character; one of 'no_change', 'background_and_margin' or 'shadow'. If 'background_and_margin', then background_color and margin parameters can be used

margin
character; margin in pixels or percentage. only used if picture_aspect is 'background_and_margin'

background_color
character; hex color. Only used if picture_aspect is 'background_and_margin'

manifest
list; list of manifest details. Should be a named list containing name; the application name (this cannot be NULL), start_url; the page that's actually added to the homescreen, display; one of 'browser' where the page is opened in a new Chrome tab or 'standalone' where the website is treated like a new application, orientation; orientation to force the screen if display is standalone, theme_color; color applied to the app when susing task switcher, existing_manifest; if there an existing manifest, it can be provided here, and on_conflict; what to do when there's a conflict between the provided and generated manifest

assets
list; list containing two TRUE/FALSE values: legacy_icon; whether to include a legacy icon, and low_resolution_icons; whether to include all documented icons or just recommended high res icons